leland helped me look through and resolve the circle initialization + helped get the updates running properly in synchCircle; after that we went through the add/remove fireflies functions
environment object got funky, so i deleted it and have been defining the functions globally. We've hard-coded the width/height parameters bc we were in the TRENCHES trying to figure out how to get the svg parameters as a value passed in based on any relevant attributes of the svg in question- if yall could help with that I'd appreciate it, otherwise I can take it to Rosenbaum or Oren. Next steps would be to get the animation set up and the lights GLOWING!!! o7 then I think we can talk  about synchronization.